[MCF-10] Make the Appointments Screen shared and use it for both roles
